Expose `build_response` method
Also remove `match` and `*args` params from callback signature
and make callback simple function instead of coroutine.
