+++ aioresponses/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from .compat import build_response
 from .core import aioresponses
 
 __version__ = '0.4.1'
