+++ aioresponses/core.py
@@ -1,29 +1,23 @@
 # -*- coding: utf-8 -*-
 import asyncio
-import json
 from collections import namedtuple
-from distutils.version import StrictVersion
 from functools import wraps
 from typing import Callable, Dict, Tuple, Union, Optional, List  # noqa
-from unittest.mock import Mock, patch
+from unittest.mock import patch
 
 from aiohttp import (
     ClientConnectionError,
-    ClientResponse,
     ClientSession,
     hdrs,
     http
 )
-from aiohttp.helpers import TimerNoop
-from multidict import CIMultiDict
 
 from .compat import (
     URL,
     Pattern,
-    stream_reader_factory,
     merge_params,
     normalize_url,
-    AIOHTTP_VERSION
+    build_response,
 )
 
 
@@ -79,66 +73,17 @@ class RequestMatch(object):
             return False
         return self.match_func(url)
 
-    def build_response(
-            self, url: 'Union[URL, str]',
-            method: str = hdrs.METH_GET,
-            status: int = 200,
-            body: str = '',
-            content_type: str = 'application/json',
-            payload: Dict = None,
-            headers: Dict = None,
-            response_class: 'ClientResponse' = None,
-            reason: Optional[str] = None) -> ClientResponse:
-        if response_class is None:
-            response_class = ClientResponse
-        if payload is not None:
-            body = json.dumps(payload)
-        if not isinstance(body, bytes):
-            body = str.encode(body)
-        kwargs = {}
-        if AIOHTTP_VERSION >= StrictVersion('3.1.0'):
-            loop = Mock()
-            loop.get_debug = Mock()
-            loop.get_debug.return_value = True
-            kwargs['request_info'] = Mock()
-            kwargs['writer'] = Mock()
-            kwargs['continue100'] = None
-            kwargs['timer'] = TimerNoop()
-            if AIOHTTP_VERSION < StrictVersion('3.3.0'):
-                kwargs['auto_decompress'] = True
-            kwargs['traces'] = []
-            kwargs['loop'] = loop
-            kwargs['session'] = None
-        _headers = CIMultiDict({hdrs.CONTENT_TYPE: content_type})
-        if headers:
-            _headers.update(headers)
-        raw_headers = self._build_raw_headers(_headers)
-        resp = response_class(method, url, **kwargs)
-        if AIOHTTP_VERSION >= StrictVersion('3.3.0'):
-            # Reified attributes
-            resp._headers = _headers
-            resp._raw_headers = raw_headers
-        else:
-            resp.headers = _headers
-            resp.raw_headers = raw_headers
-        resp.status = status
-        resp.reason = reason
-        resp.content = stream_reader_factory()
-        resp.content.feed_data(body)
-        resp.content.feed_eof()
-        return resp
-
     async def _build_response(
-            self, url: URL, *args: Tuple, **kwargs: Dict
+            self, url: URL, **kwargs: Dict
     ) -> 'Union[ClientResponse, Exception]':
         if isinstance(self.exception, Exception):
             return self.exception
-        if self.callback and callable(self.callback):
-            resp = await self.callback(self, url, *args, **kwargs)
+        if callable(self.callback):
+            resp = self.callback(url, **kwargs)
         else:
             resp = None
         if resp is None:
-            resp = self.build_response(
+            resp = build_response(
                 url=url,
                 method=self.method,
                 status=self.status,
@@ -149,17 +94,6 @@ class RequestMatch(object):
                 response_class=self.response_class,
                 reason=self.reason)
         return resp
-
-    def _build_raw_headers(self, headers: Dict) -> Tuple:
-        """
-        Convert a dict of headers to a tuple of tuples
-
-        Mimics the format of ClientResponse.
-        """
-        raw_headers = []
-        for k, v in headers.items():
-            raw_headers.append((k.encode('utf8'), v.encode('utf8')))
-        return tuple(raw_headers)
 
 
 RequestCall = namedtuple('RequestCall', ['args', 'kwargs'])
@@ -274,11 +208,11 @@ class aioresponses(object):
         ))
 
     async def match(
-            self, method: str, url: URL, *args: Tuple, **kwargs: Dict
+            self, method: str, url: URL, **kwargs: Dict
     ) -> Optional['ClientResponse']:
         for i, matcher in enumerate(self._matches):
             if matcher.match(method, url):
-                response = await matcher._build_response(url, *args, **kwargs)
+                response = await matcher._build_response(url, **kwargs)
                 break
         else:
             return None
@@ -302,7 +236,7 @@ class aioresponses(object):
                     orig_self, method, url, *args, **kwargs
                 ))
 
-        response = await self.match(method, url, *args, **kwargs)
+        response = await self.match(method, url, **kwargs)
         if response is None:
             raise ClientConnectionError(
                 'Connection refused: {} {}'.format(method, url)
